restore legal hold modal
